connect theme to the ChakraProvider and rearrange folders
