--- constants/constants.ts
@@ -1,9 +1,0 @@
-export const TRANSITION = {
-  type: "spring",
-  stiffness: 200,
-  damping: 10,
-};
-
-export const WORDS_PER_MINUTE = 200;
-
-export const NAV_ITEMS = ["Home", "About", "Portfolio", "Blog", "Contact"];
