+++ src/constants/index.ts
@@ -1,2 +1,12 @@
 export const SITE_DESCRIPTION =
   "Personal portfolio website of front-end developer Andrija Kapetanović.";
+
+export const TRANSITION = {
+  type: "spring",
+  stiffness: 200,
+  damping: 10,
+};
+
+export const WORDS_PER_MINUTE = 200;
+
+export const NAV_ITEMS = ["Home", "About", "Portfolio", "Blog", "Contact"];
